1222AM保存声音到云端

+++ Assets/Scripts/Scriptable/GameData.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "GameData", menuName = "CreateScriptable/GameData", order = 2)]
 public class GameData : ScriptableObject
 {
-    public int musicVolume;
-    public int effectVolume;
+    public float musicVolume;
+    public float effectVolume;
 
 }
